update busines rules for delete routes

+++ src/modules/deleteClassroom/DeleteClassroomService.ts
@@ -4,6 +4,11 @@ export class DeleteClassroomService {
     constructor(private classroomsRepository: IClassroomsRepository) {}
 
     async execute (id: string): Promise<void> {
+        const classroom = await this.classroomsRepository.findById(id);
+
+        if (!classroom) {
+            throw new Error('Classroom does not exist!');
+        }
         await this.classroomsRepository.delete(id);
     }
 }
