+++ src/modules/deleteCourse/DeleteCourseService.ts
@@ -1,9 +1,15 @@
-import { ICoursesRepository } from "../../repositories/ICoursesRepository";
+import { ICoursesRepository } from '../../repositories/ICoursesRepository';
 
 export class DeleteCourseService {
     constructor(private coursesRepository: ICoursesRepository) {}
 
     async execute (id: string): Promise<void> {
+        const course = await this.coursesRepository.findById(id);
+
+        if (!course) {
+            throw new Error('Course does not exist!');
+        }
+
         await this.coursesRepository.delete(id);
     }
 }
