+++ src/modules/deleteUser/DeleteUserService.ts
@@ -4,6 +4,11 @@ export class DeleteUserService {
     constructor(private usersRepository: IUsersRepository) {}
 
     async execute (id: string): Promise<void> {
+        const user = await this.usersRepository.findById(id);
+
+        if (!user) {
+            throw new Error('User does not exist!');
+        }
         await this.usersRepository.delete(id);
     }
 }
